requirements.txt
torch
torchvision
numpy
pillow
websockets
opencv-python